Add a “scydev” Vagrant box
This supersedes https://github.com/scy/devbox, which I’ll delete right
after committing this. Most of its original README went into this one.
The main difference is that I’m no longer using CoreOS as the basic
operating system, but Debian Jessie. CoreOS is great, but not as a
development machine.